[EEMW-22] | move the other components to common
