+++ src/components/application/common/age-confirmation/index.ts
